+++ src/shared/ix/map.ts
@@ -1,16 +1,8 @@
 export type MapFn<TInput, TOutput> = (input: TInput) => TOutput;
-export type AsyncMapFn<TInput, TOutput> = (input: TInput) => (TOutput | Promise<TOutput>);
 
 export function* mapForIterable<TInput, TOutput>(iter: Iterable<TInput>, transform: MapFn<TInput, TOutput>): Iterator<TOutput> {
     for (const item of iter) {
         const result = transform(item);
-        yield result;
-    }
-}
-
-export async function* mapAsyncForAsyncIterable<TInput, TOutput>(iter: Iterable<TInput>, transform: AsyncMapFn<TInput, TOutput>): AsyncIterator<TOutput> {
-    for await (const item of iter) {
-        const result: TOutput = await transform(item);
         yield result;
     }
 }
