+++ src/shared/ix/mod.ts
@@ -1,12 +1,6 @@
-export {
-    filterForIterable as filter,
-    filterAsyncForAsyncIterable as filterAsync,
-} from './filter';
-export {
-    mapForIterable as map,
-    mapAsyncForAsyncIterable as mapAsync,
-} from './map';
-export {
-    toArrayFromIterable as toArray,
-    toArrayAsyncFromAsyncIterable as toArrayAsync,
-} from './toArray';
+export { filterForIterable as filter } from './filter';
+export { filterAsyncForAsyncIterable as filterAsync } from './filter_async';
+export { mapForIterable as map } from './map';
+export { mapAsyncForAsyncIterable as mapAsync } from './map_async';
+export { toArrayFromIterable as toArray } from './to_array';
+export { toArrayAsyncFromAsyncIterable as toArrayAsync } from './to_array_async';
